pyproject.toml
[project]
name = "laptop_repair"
version = "0.1.0"
description = "laptop-repair using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.134.0,<1.0.0",
    "pysqlite3-binary == 0.5.4",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "psutil>=5.9.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.28.0"
]

[project.scripts]
laptop_repair = "laptop_repair.main:run"
run_crew = "laptop_repair.main:run"
train = "laptop_repair.main:train"
replay = "laptop_repair.main:replay"
test = "laptop_repair.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"